refactor(tools): decompose installment-vs-cash page into feature structure
Extracts 921-LOC monolith into feature folder under
app/features/tools/installment-vs-cash/:
- page.vue — shell/layout component
- useInstallmentVsCashCalculator.ts — re-export facade for model functions
- installment-vs-cash.schema.ts — Zod schema with correct DTO enum alignment
- __tests__/useInstallmentVsCashCalculator.spec.ts — type guards + domain tests

Route page app/pages/tools/installment-vs-cash.vue is now a thin wrapper.
Fixes Zod enum to match OpportunityRateTypeDto and RecommendedOptionDto contracts.
